+++ packages/foam-vscode/src/features/panels/connections.spec.ts
@@ -6,7 +6,7 @@ import {
   createNote,
   getUriInWorkspace,
 } from '../../test/test-utils-vscode';
-import { ConnectionsTreeDataProvider } from './backlinks';
+import { ConnectionsTreeDataProvider } from './connections';
 import { MapBasedMemento, toVsCodeUri } from '../../utils/vsc-utils';
 import { FoamGraph } from '../../core/model/graph';
 import {
