+++ packages/foam-vscode/src/features/panels/connections.ts
